Don't take ownership of the tensor details in the value mapper
Summary: Tidy up ownership of the ipu tensor details. Now it is only owned by torch tensor (and views of the tensor in the future). This will help simplify the logic around when tensors are alive in pytorch since the lifetime of the tensor details should be directly linked to the pytorch lifetime. This removes the need for an exlicit isAlive() funciton in the ipu tensor details

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb, georgew

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, georgew

Subscribers: georgew

Maniphest Tasks: T64272

Differential Revision: https://phabricator.sourcevertex.net/D76870