Fix `eager_simple_test.py::test_lazy_tensor` test
Summary:
Due to function caching, the expected log line is no longer output to
the log at `DEBUG` log level or higher. The compile-and-run
path is avoided entirely as the function is retrieved from the cache.
Instead, add a new log message which can be used to determine when a
PopIT function has been executed, regardless as to whether it was
compiled on demand or found in the cache and reused.

Ref T67895

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Maniphest Tasks: T67895

Differential Revision: https://phabricator.sourcevertex.net/D77237