Handle custom data in AsynchronousDataAccessor.
Summary:
Ref T71201

AsynchronousDataAccessor deals with single tensor, list of tensors and
dict of tensors data only. In the last case, it strips the keys from the
dict and exchanges only its values with the main process. If the dataset
stores data in some custom format (see Pytorch Geometric datasets),
especially, if the format is created dynamically in the runtime, there
is currently no way to re-assemble the created batch to the desired
format from the list of tensors returned from the
AsynchronousDataAccessor.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, piotrc

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, piotrc

Maniphest Tasks: T71201

Differential Revision: https://phabricator.sourcevertex.net/D78578