Handling norm op that accepts 5th argument - out Tensor
Summary:
Poptorch handler for aten::linalg_vector_norm operation reuses tensorNormHandler. To achieve that, the function had to be extended with the 5th argument processing, with assumption that the 5th arg appears in linalg_vector_norm only. It turns out, this is not true in general, as there are aten::norm versions accepting the 5th arg, but of different type (Tensor instead of Scalar). This patch enables a subset of these norm overloads.

Ref AFS-64

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, matthewha

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, matthewha

Subscribers: matthewha

Differential Revision: https://phabricator.sourcevertex.net/D79503